Use Project and HardwareReservation schemas directly in Device schema
The Device schema specifies that the `project` and `hardware_reservation` properties
are Hrefs.  With the current generator configuration, that means the Project and
HardwareReservation properties are inaccessible even if we add `include` parameters
to the device request.  (Even with fully OpenAPI spec-compliant code, the Project
and HardwareReservation properties would be accessed via Href `additionalProperties`
given the current upstream spec, so changing code generation is of minor benefit
here.)

This patches the Device schema so that the `project` and `hardware_reservation`
properties are specified as instances of the correct model schema, rather than as
Hrefs.