Fix sync job for split yaml spec
The sync GitHub Action was failing with a permissions error because
the `fetch` task in the Makefile did not specify a user.  In addition,
the command to commit the updated spec was outdated; it was still
looking for JSON files, but we now download the split spec in YAML
format.