Remove support for creating a device with a facility string
When creating a device, the `facility` property can be either a
string or a list of strings.  This construct does not work well
in generated code, and causes device creation request to serialize
incorrectly.

By patching the spec to remove support for `facility` as a single
string, we fix request serialization without losing functionality.