chore: remove combine-spec step when generating code (#102)
The OpenAPI generator is capable of generating code directly from the
split API spec (in fact, that's how we were generating the combined
spec, by running the openapi-yaml generator). This removes the combine
step from the Makefile and sync workflow in order to simplify the code
generation process.

The only side effect of this change is that the openapi-yaml generator
added values for some missing properties of the include and exclude
parameters. The properties added by openapi-yaml generator caused the
include parameters to be added as `include=<a>&include=<b>`, whereas
generating from the split spec causes include parameters to be added as
`include=<a>,<b>`, which matches existing behavior in packngo.