feat: log sunset & deprecation headers
To make it easier to adopt `metal-go` in place of `packngo`, this
copies the `packngo` logic to dump sunset & deprecation headers to
stdout.  The Equinix Metal API uses these headers to communicate
important changes and there is long-standing precedent for dumping
these headers to stdout in all of our Equinix Metal API integrations.