fix: return errors when JSON responses contain fields of wrong type (#149)
The Go generator we are using generates code that ignores errors while
unmarshalling JSON responses from the API into Go structs. This means
that if the API spec says a field will be an integer but the field in
the API response is a string, the Go SDK will return a struct with nil
values for all fields instead of returning an error.

This updates the model_simple.mustache template to make the unmarshal
function return errors when they happen rather than ignoring them.