include gateway IP reservations to enable type differentiation
There is currently no direct way to differentiate the MetalGateway
and VrfMetalGateway schemas without removing flexibility from the
MetalGateway schema.

A VrfMetalGateway is a MetalGateway with one extra property, `vrf`,
which means that an object that is a VrfMetalGateway also matches
the MetalGateway schema, because additional properties are allowed.

In metal-java, VMWare took the approach of disabling additional
properties for MetalGateway, but this is a risky approach because
it means that if the API adds new properties to MetalGateway, the
existing clients will break.

Instead of making a schema change, this uses include params in the
VRF example to include the `ip_reservation` property for each gateway.
IP reservations and VRF IP reservations are already distinguishable
based on spec/properties, so this leverages that distinction to also
determine whether a gateway is VRF or non-VRF.