fix: update spec for plan drive type to match API response (#155)
Relates to #154 (technically a fix, but I want to leave that issue open
in hopes of a better solution later)

This updates the enum items for drive types in a plan spec to change
`NVME` to `NVMe`. The original spec has `NVME`, but the API actually
uses the value `NVMe`, and the difference in case causes API responses
to be invalid according to the spec.

I tried patching in `NVMe` as an additional option, as mentioned in
https://github.com/equinix-labs/metal-go/issues/154#issuecomment-1728316194,
but that causes the generator to generate invalid code because the
constants for `NVMe` and `NVME` are both generated with the same name:

```
metal/v1/model_plan_specs_drives_inner_type.go:27:2: PLANSPECSDRIVESINNERTYPE_NVME redeclared in this block
```

We could patch the spec to remove the enum entirely for this property,
but I wanted to keep this change as small as possible.