feat: include API messages in errors
This updates the `formatErrorMessage` helper so that, if the API
responded with an Error model, the messages from that Error model
are appended to the status code & status description.  This makes
it easier for users to get visibility into why an API call failed.

I chose to leave the existing code to support rfc7807 messages in
place, just in case we ever migrate to that standard.