refactor: move all client modifications into template (#173)
We hadn't introduced custom templates when we updated `client.go` to
dump deprecation & sunset headers to match `packngo` behavior; that
change was made with a post-patch instead. Now that we are heavily using
custom templates to get the generator to spit out the code we want, we
can move all client modifications to the `client.mustache` template and
avoid potential adverse interactions between the post-patch and the
custom template.